--- packages/layers/base/app/interfaces/manage-company-name.ts
@@ -1,12 +1,0 @@
-export interface ManageCompanyNameState {
-  new: {
-    legalName: string
-    nrNumber?: string
-    actions: ActionType[]
-  }
-  old: {
-    legalName: string
-    nrNumber?: string
-    actions: ActionType[]
-  }
-}
