+++ packages/layers/base/app/interfaces/manage-components/index.ts
@@ -1,1 +1,5 @@
 export * from './manage-base-props'
+export * from './manage-company-name'
+export * from './manage-offices'
+export * from './manage-parties'
+export * from './manage-share-structure'
