+++ packages/layers/base/app/interfaces/manage-components/manage-offices.ts
@@ -1,0 +1,20 @@
+import type { ManageBaseProps } from '#business/app/interfaces'
+
+export type ManageOfficesProps = ManageBaseProps & {
+  labelOverrides?: TableLabelOverrides
+} & (
+  | {
+    variant?: 'default' | 'correct'
+    subject: string
+    modelName?: string
+    allowedActions?: ManageAllowedAction[]
+    allowAddOfficeType?: OfficeType
+  }
+  | {
+    variant: 'readonly' | 'correct-readonly'
+    subject?: never
+    modelName?: never
+    allowedActions?: never
+    allowAddOfficeType?: never
+  }
+)
