+++ packages/layers/base/app/interfaces/manage-components/manage-share-structure.ts
@@ -1,0 +1,15 @@
+import type { ManageBaseProps } from '#business/app/interfaces'
+
+export type ManageShareStructureProps = Omit<ManageBaseProps, 'tableTitle'> & {
+  labelOverrides?: TableLabelOverrides
+  tableTitle?: string
+} & (
+  | {
+    variant?: 'default' | 'correct'
+    allowedActions?: ManageAllowedAction[]
+  }
+  | {
+    variant: 'readonly' | 'correct-readonly'
+    allowedActions?: never
+  }
+)
